backend/requirements.txt
flask
flask-jwt-extended
requests
python-dotenv
gunicorn
werkzeug